+++ src/extension.ts
@@ -11,38 +11,49 @@ console.log("Extension loaded");
 export function activate(context: vscode.ExtensionContext) {
   
    const configService = new config.ConfigurationService(context);
-   //AN ERROR HERE
    const eng = new Engine(configService);
   
-  // const commands = [];
-  // commands.push({
-  //   name: "regionfolder.wrapWithRegion",
-  //   action: () => eng.wrapWithRegion()
-  // });
-  // commands.push({
-  //   name: "regionfolder.wrapWithRegionAndComment",
-  //   action: () => eng.wrapWithRegion()
-  // });
-  // commands.push({
-  //   name: "regionfolder.collapseDefault",
-  //   action: () => eng.doDefaultFold()
-  // });
-  // commands.push({
-  //   name: "regionfolder.collapseAllRegions",
-  //   action: () => eng.collapseAllRegions()
-  // });
-  // commands.push({
-  //   name: "regionfolder.deleteCurrentRegions",
-  //   action: () => eng.deleteCurrentRegion()
-  // });
+  const commands = [];
+  commands.push({
+    name: "regionfolder.wrapWithRegion",
+    action: () => eng.wrapWithRegion()
+  });
+  commands.push({
+    name: "regionfolder.wrapWithRegionAndComment",
+    action: () => eng.wrapWithRegionAndComment()
+  });
+  commands.push({
+    name: "regionfolder.collapseDefault",
+    action: () => eng.collapseAllDefaultFolds()
+  });
+  commands.push({
+    name: "regionfolder.collapseAllRegions",
+    action: () => eng.collapseAllRegions()
+  });
+  commands.push({
+    name: "regionfolder.deleteRegion",
+    action: () => eng.deleteCurrentRegion()
+  });
+  commands.push({
+    name: "regionfolder.removeCurrentRegionTags",
+    action: () => eng.removeCurrentRegionTags()
+  });
+  commands.push({
+    name: "regionfolder.selectCurrentRegion",
+    action: () => eng.selectCurrentRegion()
+  });
+  commands.push({
+    name: "regionfolder.selectCurrentRegionContents",
+    action: () => eng.selectCurrentRegionContents()
+  });
 
   
-  // for(let comm of commands){
-  //   let disp = vscode.commands.registerCommand(comm.name, comm.action);
-  //   context.subscriptions.push(disp);
-  // }
+  for(let comm of commands){
+    let disp = vscode.commands.registerCommand(comm.name, comm.action);
+    context.subscriptions.push(disp);
+  }
 
-  
+ 
 }